requirements.txt
# Core dependencies
faker==22.0.0
pandas==2.1.4
numpy==1.26.3
psycopg2-binary==2.9.9

# Database
sqlalchemy==2.0.25

# Data validation
pydantic==2.5.3

# CLI and utilities
python-dotenv==1.0.0
click==8.1.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
